HYD-5832 Job stats page allows a minimum window of 16 seconds, data slices are 10 seconds
The ui/dashboard/jobstats page shows no data for any window size of 15
seconds or less. Since data slices can be 10 seconds it may show no data
when there is data present. Data windows of at least 10 seconds, if not
less, should be supported.

This patch will update the read/write heatmap such that if the
difference between the start date and end date is < 16 seconds, x
seconds will be subtracted from the start date such that the difference
between the same end date and the new start date is exactly 16 seconds.

- Updated intel-gui to 4.4.5 in package.json
- Updated chroma-externals link

Change-Id: I5ea49e3b5c191ad369d65f92a86c2ad3ae4a9625
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/19589
Tested-by: Jenkins
Reviewed-by: Chris Gearing <chris.gearing@intel.com>