HYD-5867 On el7 await_server_boot fails because pacemaker is returning FQDN names
On el7 await_server_boot because pacemaker is using FQDN names on el7
but on el6 it uses nodenames. This means the node name used for the regex nexts
to be be varied.

The solution is to detect the version of the target using the distro_version
method in RemoteOperations.

 - Change the await_server_boot to use nodename or fqdn as appropriate.

Change-Id: I7e35fe77eba17988a994f59f32c84e5bc15d5133
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19605
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>