HYD-5854 Filesystem stats differ between EL6 and EL7
Test failure in el7 SSI tests, failure related to arithmetic which
expects defined value of statistics generated by exercising created
DNE filesystem. The results of the statistics seem to be different on
el7 than on el6.
Seeing as the test aims to detect the ability to write to the
filesystem, we should just make sure that filesystem activity has
been detected and not be concerned that the exact values are matched.
There may be hidden complexity regarding the exact measurement of
statistics.

  - modify checks for file operations (stats_open) to use 
    assertGreaterEqual rather than assertEqual. Checking for a minimum
    value will ensure that the filesystem is being exercised and will
    pass on both el6 & el7 tests.

Change-Id: Ie783ab5ef673fc9278b28b8f97a9e87f0d57a46d
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19613
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>