HYD-5671 :ElementTree.py:1245:feed:ExpatError: syntax error: line 1, column 0
Unhandled exceptions are being seen with the error message of 

   ElementTree.py:1245:feed:ExpatError: syntax error: line 1, column 0

the cause of this error is that pacemaker is being built on builders 
with the cluster-glue-libs-devel package installed. Removal of this 
package from the builders causes a pacemaker to be built that does not 
demonstrate the exception.

The removal of cluster-glue-libs-devel from the builds means that
the resulting package is a not different package that requires an
increase in the version number.

 - Update the wc release number in the pacemaker SPEC file
 - Update chroma-agent.spec to obsolete the previous version.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: Ifb1a9b7618c57cefb6a5a6ed2316d1f1cbc9fc98
Reviewed-on: http://review.whamcloud.com/19611
Tested-by: Jenkins
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>