HYD-5769 lvcreate -l 100%FREE blocks for input
The efs tests on el7 are deterministically hanging indefinitely
It was determines that it was hanging specifically on an lvcreate
that was waiting for input. The input it was wait for and answer to
was:

WARNING: zfs_member signature detected on /path/to/lv at offset 147456. Wipe it? [y/n]:

On el6, lvcreate had a default behavior to assume a 'n' and proceed without
prompting for input. However on el7 they've changed it to instead wait and ask.
The lv setup in our efs tests should be resistant to any pre-existing
signatures on the disks.

The way to avoid the prompt in el7 is to now explicitly use the
'--wipesignatures n' flag in the lvcreate to guarantee the desired
behavior.

  - Add '--wipesignatures n' to the lvcreate command in prepare_device_commands
    in test_blockdevice_lvm.py

Change-Id: I2428a9e0a4f21f8e20071934b6f79eb8e97b10ad
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/19618
Tested-by: Jenkins
Tested-by: Chroma Test User <hpdd-devops@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>