HYD-5872 Temporarily disable remaining failing tests on el7
As part of the 3.0 release, we have been getting all of the
automated tests to pass on the newly supported OS version,
el7.2. There are just a handful left, and there is great
benefit to getting the rest turned on to prevent regression.

This patch *temporarily* disables the last few failures until
they can be fixed. We will cover with manual testing until
they can be turned on on el7.2.

  - Disable failing power control tests using skip decorators
  - Skip the linux with no HA combinations in the EFS tests.

Change-Id: Ia7ab912f1baa44effa33863446fdd8ff9fae1e9a
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/19644
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Tested-by: Chris Gearing <chris.gearing@intel.com>