HYD-5886 BMC Alerts should be lowered when resource is deleted
Unable to monitor BMC alerts are raised but not lowered when the
resource is deleted. This appears to be because we may re-raise an alert
after the device is deleted.

Got into this state by:

1. creating the power control devices.
2. Since ssh keys hadn't been shared from the manager server to lotus-30
they generated alerts.
3. removed the power control devices
4. corrected the ssh keys (source of alert)
5. recreated power control devices

The alerts didn't clear and couldn't clear because the power control
devices that raised them no longer exist.

  - Remove code that may re-raise alert after deletion of resource.

Change-Id: Ib6cbbcd296c97918c6d1e7ac20552ac2300cffdc
Signed-off-by: Joe Grund <joe.grund@intel.com>
Reviewed-on: http://review.whamcloud.com/20037
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>