HYD-5895 Use "best run" instead of "latest run" in test aggregation
Currently the test reporting aggregation script takes the results
of the most recent run for each test suite. This can lead to some
frustrating situations such as two people both kick off re-runs,
the first one passes and marks the review with a check, but then
the second one hits an intermittent failure and undoes the check.
Even if you abort the second one, it is still counted. Also, now
that we have a matrix of distros to test, it can be difficult to
get both distros to pass at once when intermittent failures are
frequent.

This patch changes the test aggregation to instead pick the "best"
test run. This means that it will take the most recent passing run,
or the most recent failing run if none are passing. This should make
it easier to get past test while still making sure that every test
suite passes at least once.

  - Modify chroma-manager/tests/framework/test_reporting/aggregate_test_results.py
    to select the "best run" rather than the "most recent run".
  - Also fixes a potential false positive situation. If any of the selected runs
    fail, the overall report needs to be a fail.
  - Makes some improvements to logging to be able to see which runs are
    considered, and which ones are ultimately selected.
  - Is smart enough to exclude currently running tests and tests that have
    fallen out of history.
  - Upgrade jenkinsapi to 0.2.28 to get fix for bug in is_good and
    efficiency improvements. Does not upgrade all the way to 0.2.29 as
    that is incompatible with our current version of jenkins.

Change-Id: I2990e248907b5d84736913e1b289222a36285680
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/20658
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>