HYD-5899 Test console output format should be unified
When we run tests we receive the output from unittest runner on
the console and in logs. This output is either the default which
includes name of test, class and module and the result, or if
docstring is present, a truncated version of the docstring.

Currently we have a mix of both formats in our output, it would
be best to just have the default format.

 - Set self.descriptions to 0 before calling _makeResult()

Change-Id: I0fbc3bea277cd56f31a30fd4ca8a846b227e8d0d
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/19794
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>