HYD-6155 Default to all search options for rest endpoints and provide enumeration
To make the GUI parser more effective we should enable all possible options for
filters rather than a subset. Also we need to provide the enumeration values
for enumerated types in the schema.

For a first ticket add enumeration types and full filtering to the alert and
the log endpoint. Additional endpoints can easily be added as over time.

Add new property to the ApiField class to be a list of enumeration types. This
is done using a monkey patch because we cannot easily change the classes
that tastypie is inheriting from.

As an example message_class in the log schema now contains

message_class: {
    blank: false,
    default: "No default provided.",
    enumeration: [
        "NORMAL",
        "LUSTRE",
        "LUSTRE_ERROR",
        "COPYTOOL",
        "COPYTOOL_ERROR"
    ],
    help_text: "Unicode string.  One of ['NORMAL', 'LUSTRE', 'LUSTRE_ERROR',
'COPYTOOL', 'COPYTOOL_ERROR']",
    nullable: false,
    readonly: false,
    type: "string",
    unique: false
}

 - Use enumeration type and expanded alerts for alert endpoint.
 - Use enumeration type and expanded alerts for log endpoint.

Change-Id: Idcb387384b509b619d8ae9e5e8b55628d135a22c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19797
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>