HYD-5909 fence_chroma list command returns fqdns on el7
In test;
(remote_operations.py:RealRemoteOperations.get_fence_nodes_list), the
command "fence_chroma -o list" is used to query pacemaker for
'fenceable' nodes and return a list.

The list consists of short names/nodenames on el6 platform versions of
pacemaker/fence_chroma but fqdns on el7 platforms.

The solution is to always return fqdns and transform nodenames to
fqdns when returned from the list command.

  - modify get_fence_nodes_list to always return a list of fqdns
  - change usages of get_fence_nodes_list to expect a list of fqdns

Change-Id: I4a4d0e8e090a4e52fcbbfc2f2b981035f7ece226
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/19795
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>