HYD-5933 anyjson fails to build
Any json is failing to build under some not understood
circumstances. The cause at a superficial level is that
although the build process tries to install.

python-anyjson-0.3.3-1.el6.noarch.rpm

the build actually creates

python-anyjson-0.3.3-1.noarch.rpm

The root cause of this is somewhere in the general build system but
in the case of anyjson there is no reason to build it as
we can just use the pre-built version from the epel repo.

This tickets adds anyjson to the epel dependencies and removes
the anyjson directory from the chroma-dependencies.

A final step is the ability to install packages into the build
environment from the epel dependencies

 - Remove anyjson from chroma-dependencies.
 - Add anyjson to chroma-dependencies/epel
 - Remove python-anyjson-0.3.3-1.el6.src.rpm
 - Remove python-anyjson-0.3.3-3.el7.src.rpm
 - Add python-anyjson-0.3.3-1.el6.noarch.rpm
 - Add python-anyjson-0.3.3-3.el7.noarch.rpm
 - Add install_pkg and uninstall_pkg to the epel Makefile

Change-Id: Ia72333112ca3b7ae3ae7f376446868d95cf240e1
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/19998
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>