HYD-2889 Timed out waiting for host to come back online
This is a debug ticket for the generic failure described above. It
is designed to provide a bit more detail as to what is actually happening
during the failure. The additional detail is to check if the host is
actually contactable - is it actually alive or just not reported alive.

In the future we can expand to provide even more highly valuable and
informative information.

 - Extend await_server_boot to report additional information if the node
   does not return in TEST_TIMEOUT period.
 - Enable fetch_help to just send the exception message to the user if
   none is provided

Change-Id: I7f25fd1709d9628d3715943e02d50bcb53070548
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20113
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>