HYD-5996 ClientConnectEvent and SyslogEvent contain 0 as a default for a string
ClientConnectEvent and SyslogEvent contain 0 as a default for a string which
creates an exception in the event that something tries to create a ClientConnectEvent.

These errors where found by Abiola improving the behaviour and by inference the
test for this class in ticket HYD-5951.

This ticket is a spot fix for the failure to define the correct type failure and
toughen the code validation to prevent similar failures happening again.

 - Make default for ClientConnectEvent '' instead of 0
 - Make default for SyslogEvent '' instead of 0
 - Add validation to VariantDescriptor to make it more robust against invalid
   possibilites.

Change-Id: If592188ea014306b1da2efb8a569027ac3ea9473
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20134
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>