HYD-6035 GUI Version bumps should only run the service test suite
When the GUI is bumped we should selectively run tests. The tests required
for GUI are the Service Tests.

Add functionality that can be extended to selectively remove test suites
when only a subset of files change. For the initial version only remove test
suites when the only change is to file 'chroma-manager/ui-modules/package.json'

 - Add bash function to skip a test whilst returning valid results
 - Add bash function to check if the patch is only a gui bump
 - Add a select_auto_pass function which can be extended over time to
   selectively skip tests suites depending on the changed files

Change-Id: If1a8345513f6e931f5cd837218ab5417d2cfc9d6
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20139
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/21249