HYD-5882 pcs --force cluster node remove throws an error.
Removing corosync configuration with pcs in parallel is prone to race
conditions because both nodes may remove themselves whilst still considering
that another node is active. This leads to an error "Error: Unable to update
any nodes" because the action of removing itself causes it to become the
last node.

Serialize the removal of nodes from the corosync cluster to ensure that race
conditions do not occur with cluster removal. Generally serialization is bad
but in this case we have only 2 node clusters and the operation is relatively
quick. We also have no choice because pcs obviously doesn't handle the case
of the last 2 nodes simultaneously but independently being removed from the
cluster. If we increase the size of the clusters then we might want to change
the mechanism but we would need to be very careful in our test before doing
so. At present however a safe as houses approach probably makes sense.

 - Add a lock around the unconfigure_corosync2 so it is executed serially
   across all the nodes in a HA cluster (2 nodes)

Change-Id: I96d13ec140ed7ffa0581c9f44585018ec6c249f8
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20437
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
(cherry picked from commit 9c5d653aae153a7110d241a6d7154e8cad4e318b)
Reviewed-on: http://review.whamcloud.com/20438