HYD-5826 plugin_runner crashes when device busy during stat
There are certain actions that can happen on a system that
cause os.stat to raise an error even though a device actually
exists. For example, running a partprobe causes the partition
table to reload. So if an os.stat hits at the exact moment
partprobe is reloading the partition table, it will raise an
error. This error is becomes an exception later when we try
to look the drive up and crashes the plugin_runner.

This highlights two ways we could be more defensive. One is
to retry os.stat on the devices for a short window to allow
us to see the device when it is just temporarily unavailable.
Secondly, we should cope with getting back a None for a drive
rather than crashing the plugin_runner.

  - Add a short retry loop to device_helper.py when we call
    os.stat on the device paths.
  - DRY up a common pattern of returning an empty list
    when a path isn't found when looking up the major_minor
    (_paths_to_major_minors).
  - Make sure calls to _dev_major_minor handle getting
    None as a response. References in:
      chroma_agent/device_plugins/linux.py
      chroma_agent/device_plugins/linux_components/device_helper.py
      chroma_agent/device_plugins/linux_components/zfs.py
  - Add unit tests for _dev_major_minor and the new
    _paths_to_major_minors.

Change-Id: I98e605d02e7f6c42040c17fe4d85700f257b7da1
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/20653
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>