HYD-6045 Make BlockDevice and FileSystem classes abstract
BlockDevice and FileSystem classes currently behave as though they are
abstract. A more standard Python method of emulating abstract classes
is to use the abc Python module.

These classes abstract the underlying functionality of the
backend file system types e.g. ldiskfs and zfs.

Changing these to better emulation of an abstract class also uncovered
a bug which was that whilst mkfs_options was abstract FileSystemZfs did
not implement it.

 - Inherit FileSystem and BlockDevice from abc
 - Inherit TestFileSystem and TestBlockSystem from abc
 - Implement mkfs_options in FilesystemZFS and return an empty list

Change-Id: I5f54289fc72f2ecee056edccc7aa85f1f1ae9204
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20703
Tested-by: Jenkins
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>