HYD-6116 Django_celery is blacklisted and needs to be removed
Django_celery is blacklisted and needs to be removed. The only use of the
celery libraries is to send emails for alerts.

A simpler solution is to run a thread to send alerts and then make that thread
pend on an event that can be raised whenever an alert is created. This is
simpler and actually more efficient than the poll every 30 seconds made today.

 - Remove django_celery and celery libraries.
 - Remove the worker and worker_multiple processes as they are not required now
 - Encapsulate the email code into a class that an be instantiated when the job
   scheduler starts and then remain monitoring db changes to look for changes to
   alerts.

Change-Id: Ia5d148b90a03e10900ef3709db7c5c41609da7c5
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20782
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>