HYD-863 Integration Test Starting and Stopping Targets
Integration test coverage to test stopping and starting
various types of targets and ensuring that they are
genuinely stopped and started according to pacemaker
and verifying its status using the cluster resource manage
(crm_resource) command on the lustre servers to see if
it is currently considered running or not.

This was achieved by setting up a filesystem and stopping
various targets OST, MGT , MDT), checking the state was
stopped then start them then check again to make sure they
are running.

   -Create filesystem
   -start and stop various targets
   -check that they are running when started and not when stopped

Change-Id: I1972d6fa441b359f95afbc325a7a26c4b0fd314c
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/20867
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>