HYD-6115 tests.integration.shared_storage_configuration.test_ntp.TestNtpSync.test_ntp_sync
This test sets the clocks 8 hours out of sync before installing the nodes.
During the install on pacemaker will sometimes appear to fail to start. However
looking at the logs the wait for pacemaker to start seems short this is it would
seem by the wait function actually using wallclock time as a timeout and so it
time changes it gets it wrong.

Make the wait function wallclock time agnostic and as the function appears twice
move it to common code.

 - Move the wait function from utils/__init__.py and chroma_agent/utils.py
   to chroma-common/lib/utils.py
 - Change the function to not make use of wallclock time hence making it walk
   clock time changes agnostic.
 - Remove one of the unsed options to simplify the function even further.
 - Change the callers to use the moved and modified function.

Change-Id: Ifae9be0b3c0e5335faf1060db4d7328516bd5d21
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20825
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>