HYD-6114 Reporting job only seeing latest when consecutive runs
Our test aggregation script has logic to look back at past runs
to determine which is the "best" run. However, there is a defect
in the logic that puts together the list of all of the runs such
that if there are consecutive runs in Jenkins, it is only
considering the last one in the range. If they are non-consecutive
they are already considered properly.

We should instead include all of the consecutive runs in the list
of runs to be evaluated.

  - Change the api call that gets the fingerprints to include the
    start of ranges.
  - When creating the list of runs, include all runs between the
    'start' and 'end' for the fingerprint.

Change-Id: I12e133fdeefbfd3a9f2d0628452d12df1db276b6
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/20832
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>