HYD-6020 Add zfs/spl binaries to IEEL zfs bundle
In order to install zfs and dependencies onto an agent we should be
able to install the user-space tools and use dkms packages
to create relevant modules. This should now be possible due to
agreement from the PDT with regard to shipping the ZFS libs/binaries.
create_installer scripts will still be used for monitor only or
non-IML IEEL installations.

  - zfs, spl binaries and zfs library dependencies included in
    chroma-dependencies/zfs and chroma-bundles/zfs
  - code tidy-up resulting in no behavioural change other than
    a string testing fix for zfs "make download" target

Change-Id: I8207eacbba1661b9d797695c5d65d42a02d1ddd1
Signed-off-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-on: http://review.whamcloud.com/20898
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>