HYD-864 Integration Test Starting and Stopping a Filesystem
Integration test coverage to test stopping and starting
the filesystem  and ensuring that the filesystem is
genuinely stopped and started according to pacemaker.
Also to check that the targets OST/MDT is unmounted when
the filesystem is stopped and mounted when the filesystem
is availalabe, verifying its status using the cluster
resource manage (crm_resource) command on the lustre
servers to see if it is currently considered running or not.

This was achieved by setting up a filesystem then check if its
current available and do the checks then stop the filesystem
then do the checks then start up the filsystem again.

   -Create filesystem.
   -start and stop filesystem.
   -check that the targets are mounted when started and not
    when stopped.
   -check filesystem has stopped/started.

Change-Id: I64a07c4b894ee07f9d76b9df9b8fe046a320aaaf
Signed-off-by: Abiola Saraki <abiola.saraki@intel.com>
Reviewed-on: http://review.whamcloud.com/20892
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>