HYD-6164 Error migrating from IEEL 2.4.1.1 to IEEL 3.0.0.0
The upgrade to IEEL 3.0.0.0 can fail because some of the migrations
of alerts was incomplete.  This lead to assertions and errors in the
migrations. A knock on effect was that because of the use of nested
transactions the upgrade could not be retried because the migration
was part complete.

The migration for StorageResourceOffline need to be added, plus
the migration for LearnEvent had a misnamed field.

The migration routine also needed tweaking for fields where there
were  no custom fields but record_type still required filling to the
correct name rather than just being a copy of alert_type

 - Add migration for StorageResourceOffline
 - Rename LearnEvent.learned_item_type to LearnEvent.learned_item_type_id
   Ensure that record_type is always a real type and not a calculated type.

Change-Id: I285da81af517059f76f9428784a4130963027044
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20943
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>