HYD-5523 HYD-5524 Support EL 6.8
EL 6.8 was released in May 2016. Now that we've landed the fixes
necessary for IML to work on el6.8, we need to make the installer
allow IML to be installed on el6.8 nodes. In order to actually test
this, our tests scripts also need to not hardcode references to
just el7 and instead allow for variable test distributions.

To do this, we need to add 6.8 to the installation whitelist,
and fix a hardcoded reference in a test script to el6.7.

 - update install to include 6.8
 - update references to 6.7 in test script to variable.

Signed-off-by: Brian J. Murrell <brian.murrell@intel.com>
Change-Id: I41fb638ee71171ec6a93f705e327525eda6b9e23
Reviewed-on: http://review.whamcloud.com/19950
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>