HYD-6127 Actively move target devices between nodes during operations
An oversight in the managed ZFS implementation is that whilst HA takes 
care of moving target-devices about (with the aid of the agent) once 
the target is active it does not take care of it before the target is
active and so operations like formatting may failed because  the target 
might well be in the active on the wrong node.

A mechanism is required to move the device between nodes before some 
operations such as format. This operation will need to make sure all
nodes have the device 'exported' before then making sure that the device
to be operated on is 'imported' on the active device.

The list of other nodes will be all the nodes in the VolumeNodes list
for the device, this may include nodes that are not even be in the same
HA cluster.

By using a step to carry out this action the step can be used in jobs
other than format device job (FormatTargetJob) in the future.

The new MakeTargetActiveStep issues the 'export' operations in parallel
across all the inactive nodes followed by the 'import' operation on the
active node.

Two new agent commands import_target and export_target are added to 
facilitate the operation.

The information required to carry out HA operations is changed with
target_name being removed and device_type being added.

The BlockDevice class has been extended to include an import_ and export
method, with suitable methods being added to BlockDeviceZFS.

The FileSystem class was simplified with the parameters to mount and
unmount being changed to just the mount_path and no-parameters
respectively.

An emulation of threading is added for unit tests because when running 
unit tests every test is within a transaction and so if you kick off 
another thread you will not see	any of the database changes that have
occurred.

Appropriate and required improvements were made to testing.

 - Add MakeTargetActiveStep to export and import appropriately.
 - Add MakeTargetActiveStep to FormatTargetJob.
 - Add import_target and export_target to chroma agent.
 - Add import_ and export to BlockDevice class.
 - Simplify mount and unmount to the FileSystem class.
 - Add tests for new code.
 - Improve debug of issues with errors in jobs.
 - Allows emulation of multithreading in unit tests.

Change-Id: I2ca8e7107c56622f6b767fd1cab581d52a6d41c4
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/20971
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Will Johnson <william.c.johnson@intel.com>