HYD-6165 Rest interface selective fields not working as expected
HYD-4821 added a feature whereby the expansion of fields in the rest interface
could be controlled using a dehydrate__{field_name} entry in the request.
However it has become apparent that {field_name} is really field type which
for almost every case is as expected because host is generally host. But for
the filesystem endpoint mgt, mdts and osts could not be expanded using for
example dehydrate_mgt instead dehydreate_target needed to be used because 
target is the type.

Change dehydrate_related to use the name of the references field rather than
the type this does mean that the __{field_name} will be the exact field name
being plural for lists and singular for singulars. hosts [] vs host etc.

 - Update dehydrate_related to use instance_name rather than the instance
   type
 - Updates tests to reflect the changed names

Change-Id: I6a6af6f90c38913a94212ecf8229c4817a373c6c
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21694
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Abiola Saraki <abiola.saraki@intel.com>