HYD-6173 Replace nconf with a new configuration method in view-server and realtime HYD-6172 Replace bunyan with custom logger
nconf has been marked as having dependencies that are black listed.
Instances of nconf should be removed from the base.html and
backend_error.html files.

This patch will remove instances of conf.get('VAR') and replace them
with conf.VAR as it will now be using a POJO.

Some bunyan deps were blacklisted.

Replace bunyan with a custom logger.

The log format is rather simple, here is an example from the realtime service:

{
  "name":"realtime",
  "hostname":"lotus-30vm13",
  "pid":19872,
  "level":30,
  "sock": {
    "id": "/#z7g2SQ5zGkapWpjBAAAK"
  },
  "msg": "stream ended",
  "time": "2016-07-06T19:12:29.135Z",
  "v":0
}

Basically it's just a JSON object that is then searchable by bunyan-cli.
If we create a logger that uses the same format, we can still use bunyan-cli to search / filter records.

  - Update deps for removed items.
  - Update conf strings.
  - Remove inlined help in view server's conf.json.

Change-Id: I6275db2564d6b5213edb2f5ef833224b6f78da88
Signed-off-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-on: http://review.whamcloud.com/21222
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Joe Grund <joe.grund@intel.com>
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>