HYD-5625 AssertionError: CorosyncNoPeersAlert, CorosyncNoPeersAlert
When we delete a DeletableObject we lower all Alerts associated with it.

This creates a race because an Alert could be raised as the DeletableObject
(generally a Host) is deleted. Leading to an Alert being in the raised condition
on a deleted object.

If we really deleted the object (host) then the database would deal with the
issue at the time, either by dealing with it or raising an exception. Our use
of not_deleted means we just end up with an inconsistent database!

So what is the solution that allows us to add the integrity back into the
system.

  1. We could scan every so often and sort it out!
  2. We could hide the problem in the rest interface by adding that the
     alert_item must not be deleted to the query.
  3. Work out how to do proper integrity, see if we can check at the point the
     Alert is raised.

3 is the correct option and requires (other options are welcome) locking the
table being deleted when we lower any Alerts connected to the object and then
locking the table when we create an Alert - and when we create an Alert we must
not raise it if the object is actually deleted. However after two days of
trying to make solution 3 work I had to give up because I could not
get a solution to work reliably.

So I came up with an alternative close to option 2 which is in the rest
interface. It is a not very nice solution to HYD-5625. The problem with
HYD-5625 is that an alert_item can be deleted at the same time as another
thread/process is raising an alert on that item. Because the link between
the two is computed AND because we don't really delete the alert item
generally but mark it as deleted the database cannot use simple integrity
to prevent this happening. So I went for a solution which is to mark any active
alert that has a deleted item as inactive when it fetches from the rest interface
if it is found the alert is marked as in active. This fix is definitely a
fix not a prevention, but working with what we have I've not seen a prevention
that is possible without a major rewrite for an occasional issue.

Additionally a call back is used so that a developer does not need to remember
to mark an alert inactive when doing a delete of an item which might be an
alert item. This should lead to more robust code.

 - Add fixup code to rest interface endpoint.
 - Add callback to delete to that any item which is an alert_item will cause
   the alert to become inactive when the alert_item is deleted.

Change-Id: I5ac872f45dd6cf5f9181481d7025e89b3df0d832
Signed-off-by: Chris Gearing <chris.gearing@intel.com>
Reviewed-on: http://review.whamcloud.com/21243
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Tom Nabarro <tom.nabarro@intel.com>
Reviewed-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-by: Joe Grund <joe.grund@intel.com>