HYD-6176 Remove el6 from EE 3.x
Product management has decided to remove support for el6
support in the 3.x product line. This means we need to not
build or test against el6 any longer.

We need to remove el6 from the builds, skip test on el6 slaves,
and remove the el6 managed profile. We are only doing a
superficial removal at this time in case the decision is
reversed once we see a response in the field. Full excision of
el6 codepaths will be for another day.

  - delete base_managed.profile
  - remove base_managed.profile from chroma-bundles/Makefile
  - remove el6 from Makefile.distro as a recognized distro
    to prevent accidental building.
  - remove 6.8 as a supported distro in selective_auto_pass
    so we stop testing on el6.

Change-Id: Ia5ece3203d3861014153896fb3a946b2cb2d6a19
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21250
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Will Johnson <william.c.johnson@intel.com>
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>