HYD-6194 Upgrade test breaks on expired metadata on el7
The upgrade test code enables and disables the epel repo as part
of its testing. However, the regex that parses the exact repo
name if broken on el7 due to a leading '!' on the repo name. This
exclamation prepends the repo name when the repo metadata is
expired on el7, and the repo data is always expired at this point
in the test because the previous line replaces the repo file.

We should clean the repodata after replacing the repo file to
clean our the old repo metadata.

  - The 'yum clean all' that is currently only happening on RHEL
    really needs to happen on both distros as the repo data will
    always be expired after replacing the repo file.

Change-Id: Ib14a4309e84da286e87368cb2cd240f29c40a8da
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21415
Tested-by: Jenkins
Reviewed-by: Brian J. Murrell <brian.murrell@intel.com>
Tested-by: Chroma Test User
Reviewed-by: Chris Gearing <chris.gearing@intel.com>