HYD-6199 Re-export of TEST_DISTRO_VERSION has incorrect scope
There is a line in the upgrade_os.sh test script that re-exports
the TEST_DISTRO_VERSION variable. However, this is called outside
of the function, which means it actually executes when the script
is sourced, not after the upgrade is done as intended. This results
in setting TEST_DISTRO_VERSION to '' instead of the new version.

Move the re-export of TEST_DISTRO_VERSION to inside of the
upgrade_os function so it is actually executed after the os upgrade
is completed.

  - move TEST_DISTRO_VERSION export into upgrade_os in
    chroma-manager/tests/framework/integration/utils/upgrade_os.sh

Change-Id: Iafeb9e77fa109e297b4a714a849c301b195d1558
Signed-off-by: Kelsey Prantis <kelsey.prantis@intel.com>
Reviewed-on: http://review.whamcloud.com/21390
Tested-by: Jenkins
Tested-by: Chroma Test User
Reviewed-by: Wayne Kenneth Seymour III <wayne.k.seymour.iii@intel.com>
Reviewed-by: Chris Gearing <chris.gearing@intel.com>